fix XLA metadata for primitives with many args
